Update docs.yml
updating this as it appears to not be doing as intended, namely constructing a list of changed files and checking for the presence of certain docs source files in there. 

upon a little investigation with local clones, it seems that the possible reason for this (still not 100% sure) may be the laughably minor use of three dots rather that two in the `.yml` file. 

so this PR fixes that